Completed CamSummary backend

+++ Backend/src/main/java/com/uwc_cam_champion/uwc_cam_champion/camsummary/CamSummaryRepository.java
@@ -2,6 +2,6 @@ package com.uwc_cam_champion.uwc_cam_champion.camsummary;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface  CamSummaryRepository extends JpaRepository<Cam,Long> {
+public interface  CamSummaryRepository extends JpaRepository<Cam,Integer> {
     
 }
